% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msLandscape_convertToEEMSDiffs.R
\name{msLandscape_convertToEEMSDiffs}
\alias{msLandscape_convertToEEMSDiffs}
\title{msLandscape_convertToEEMSDiffs is a utility function to convert \emph{ms} output for use with \emph{EEMS}}
\usage{
msLandscape_convertToEEMSDiffs(dirWith_4eemsdiffs_files)
}
\arguments{
\item{dirWith_4eemsdiffs_files}{The path to the directory containing the .4eemsdiffs files to convert to .diffs files for 
EEMS using pairwise genetic differences between each individual at each locus. NOTE - the Python msLandscape script to 
convert ms output into SpaceMix input and EEMS .4eemsdiffs files needs to be run before this script.}
}
\value{
None
}
\description{
This function is the second processing step required to convert raw \emph{ms} simulation output into the
pairwise genetic differences format required by \emph{EEMS}. The accompanying Python script 
ghostLands_msOutputConvertFor_SpaceMix_EEMS_for_msLandscape.py takes the raw \emph{ms} simulation results
and converts them to the intermediate '.4eemsdiffs' files. This msLandscape_convertToEEMSDiffs
function then converts those '4eemsdiffs' files to '.diffs' files for use with EEMS.
}
