% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msLandscape_networkPlotter.R
\name{msLandscape_networkPlotter}
\alias{msLandscape_networkPlotter}
\title{Plot the landscape configuration (populations and their migration connections) in an \emph{ms} flag file.}
\usage{
msLandscape_networkPlotter(msFlagFileName, addPopnLabels = TRUE,
  savePlotToFile = FALSE, plotWidth = 8, plotHeight = 8,
  outputFileName = "msLandscape_NetworkPlot.pdf")
}
\arguments{
\item{msFlagFileName}{is the name (with the path if the file is not in the working directory) of the ms flag file to plot.}

\item{addPopnLabels}{boolean (TRUE/FALSE); default TRUE. Whether population labels (numbers) should be added to each population (point)
on the network plot. This is useful when specifying the populations to edit (either to remove or to change the number
of individuals sampled per population), and is useful too for determining the orientation of the landscape when the
network plotter rotates or reflects it compared to what is expected.}

\item{savePlotToFile}{boolean (TRUE/FALSE); default FALSE. Chooses whether the plot is saved directly to a file (pdf), or 
printed to the screen instead.}

\item{plotWidth}{define the width of the saved .pdf plot (in inches). Only used if savePlotToFile is TRUE.}

\item{plotHeight}{define the height of the saved .pdf plot (in inches). Only used if savePlotToFile is TRUE.}

\item{outputFileName}{specify the name (and path if necessary) of the output .pdf plot file. Only used if savePlotToFile is TRUE.}
}
\value{
None
}
\description{
This uses a network graph to automatically represent the landscape configuration specified in the input \emph{ms} flag file.
In this network graph, the nodes represent populations on the landscape and the edges represent 
pairwise migration connections between populations.  Node size is proportional to the number of individuals
sampled from each population, and edge width is proportional to the migration rate specified for each pair
of populations. Populations that only have one specified migration connection in the ms flag file (and therefore
uni-directional migration) are connected with a dashed line instead of a solid line.
The network plotting engine is from the igraph package. Importantly the network plot uses layout 'layout_with_kk';
 other layouts did not work well and this one works extremely well.
}
